Add Stoppable interface
Elevator, ECS and PassengerGenerator share the same logic. They are
started and stopped the same way, so it makes sense for them to
implement the same interface. This also makes it easier to make
control them in the Main method with less code (e.g.
`runnables.forEach(Stoppable::stop;`)

Changes also include:
 - Stopping all threads if one of them fails, using
ScheduledExecutorService
 - Making shouldRun volatile. Makes no difference now, but it's a good
 practice