Improve concurrency and refactor
Problem: Synchronizing on Elevator.movementDirection resulted in ECS thread
spending a lot of time blocked. This was expected and desired, but not
to the degree where ECS stops reading buttons and looking for
destinations, until direction can change.
Solution: Remove synchronization in favor of checking Elevator.isMoving
value in direction setter, and adding atomicity to ElevatorDirection.

Also:
 - Made a bunch of shared resources Atomic or volatile
 - Rearranged fields project wide
 - Replaced some occurrences of String concatenation by "+" with
 String.format() for better performance